+++ dymension/libs/kaspa/lib/relayer/src/confirm/mod.rs
@@ -1,2 +1,4 @@
 pub mod confirmation;
 pub use confirmation::*;
+mod confirmation_test;
+
